backend/pyproject.toml
[build-system]
requires = ["setuptools>=69", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "kuberatreasury"
version = "0.1.0"
requires-python = ">=3.12"
dependencies = [
    "fastapi>=0.111",
    "uvicorn[standard]>=0.29",
    "sqlalchemy>=2.0",
    "alembic>=1.13",
    "psycopg[binary]>=3.1",
    "pydantic>=2.7",
    "pydantic-settings>=2.2",
    "passlib[bcrypt]>=1.7",
    "bcrypt>=4.1",
    "python-jose[cryptography]>=3.3",
    "pyotp>=2.9",
    "httpx>=0.27",
    "python-multipart>=0.0.9",
    "email-validator>=2.2",
    "reportlab>=4.1",
    "cryptography>=42",
    "babel>=2.14",
]

[project.optional-dependencies]
dev = [
    "pytest>=8",
    "pytest-asyncio>=0.23",
    "pytest-cov>=5",
    "httpx>=0.27",
    "factory-boy>=3.3",
    "freezegun>=1.4",
    "aiosqlite>=0.20",
    "ruff>=0.6",
    "mypy>=1.11",
    "pip-audit>=2.7",
]

[tool.pytest.ini_options]
asyncio_mode = "auto"
testpaths = ["tests"]
addopts = "--cov=app --cov-report=term-missing --cov-fail-under=80"

[tool.coverage.run]
omit = ["tests/*", "alembic/*"]
